% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_knn_result.R
\name{predict_result}
\alias{predict_result}
\title{KNN Predict (No SHAP)}
\usage{
predict_result(new_data)
}
\arguments{
\item{new_data}{data.frame，包含列: ABCB5, AC093642.5, AF146191.4, ANO3,
CTD.2231H16.1, PRKG1, RASA3, VIPR2}
}
\value{
List: 包含 pred_prob, pred_class
}
\description{
给定 8 个特征的数值输入，输出该样本预测概率与分类结果。
}
